% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contrast.R
\name{check_contrast}
\alias{check_contrast}
\title{Color Contrast Accessibility Report}
\usage{
check_contrast(fg_col, bg_col)
}
\arguments{
\item{fg_col}{the Foreground Color}

\item{bg_col}{the Background Color}
}
\value{
Color Contrast Report
}
\description{
returns a report from the Contrast Checker API about color contrast for accessibility
}
\examples{

check_contrast(fg_col = "#21EA06", bg_col = "#483D3D")


}
