% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contrast_raw.R
\name{check_contrast_raw}
\alias{check_contrast_raw}
\title{Color Contrast Accessibility Report in a Raw Format}
\usage{
check_contrast_raw(fg_col, bg_col)
}
\arguments{
\item{fg_col}{the Foreground Color}

\item{bg_col}{the Background Color}
}
\value{
Color Contrast Report in a raw format
}
\description{
returns a report from the Contrast Checker API about color contrast for accessibility in a list format
so that the information provided can be extracted and piped into other functions.
}
\examples{

check_contrast_raw(fg_col = "#21EA06", bg_col = "#483D3D")


}
