% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description.R
\name{describe_using}
\alias{describe_using}
\title{Describe an HTML element by another one}
\usage{
describe_using(element, descID)
}
\arguments{
\item{element}{the HTML element to describe}

\item{descID}{the ID of the HTML that will be used to describe the 'element'}
}
\value{
an HTML element described by another HTML element
}
\description{
Describe an HTML element by another one
}
\examples{
if (interactive()) {
ui <- fluidPage(
  # use Tab and Shift + Tab to navigate between the buttons
  # and see the difference

  actionButton(inputId = "inp1",
               label = "button"),
  actionButton(inputId = "inp2",
               label = "button"),

  actionButton(inputId = "inp3",
               label = "button") \%>\%
    describe_using(descID = "descriptor"), # it's happening here

  actionButton(inputId = "inp4",
               label = "button"),
  actionButton(inputId = "inp5",
               label = "button"),

  div(id = "descriptor",
      "button 3 is just amazing! click on it!!!")
)

server <- function(input, output, session){}

shinyApp(ui, server)

}
}
