% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_invisible.R
\name{make_invisible}
\alias{make_invisible}
\title{Make an element invisible}
\usage{
make_invisible(element)
}
\arguments{
\item{element}{the element to make invisible}
}
\value{
an invisible HTML element
}
\description{
Make an element invisible so that it can only be read
by screen readers
}
