% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_skiplinks.R
\name{to_skiplinks}
\alias{to_skiplinks}
\title{Title}
\usage{
to_skiplinks(element, skip_to, bg_color = "#212121", col = "#F2F2F2")
}
\arguments{
\item{skip_to}{the html element to skip to}

\item{text}{text displayed in the anchor}
}
\value{

}
\description{
Title
}
