% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_invisible_anchor.R
\name{create_invisible_anchor}
\alias{create_invisible_anchor}
\title{Make an element invisible}
\usage{
create_invisible_anchor(id, text, href = NULL)
}
\arguments{
\item{element}{the element to make invisible}
}
\value{
an invisible HTML element
}
\description{
Make an element invisible so that it can only be read
by screen readers
}
