% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description.R
\name{describe_using}
\alias{describe_using}
\title{Describe an HTML element by another one}
\usage{
describe_using(element, descID)
}
\arguments{
\item{element}{the HTML element to describe}

\item{descID}{the ID of the HTML that will be used to describe the 'element'}
}
\value{
an HTML element described by another HTML element
}
\description{
Describe an HTML element by another one
}
\examples{



}
