% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description.R
\name{describe_using}
\alias{describe_using}
\title{Describe an HTML element by another one}
\usage{
describe_using(element, descID)
}
\arguments{
\item{element}{the HTML element to describe}

\item{descID}{one or a vector of many HTML elements' <IDs>
that will be used to describe the 'element' parameter}
}
\value{
an HTML element described by another HTML element
}
\description{
Describe an HTML element by another one
}
\examples{
if (interactive()) {
  library(shiny)
  library(magrittr)

  ui <- fluidPage(

    h2("Using a screen reader
        hit Tab and Shift + Tab to
        navigate between the buttons
        and stop at button 3 to see the difference"),

    actionButton(
      inputId = "inp1",
      label = "button 1"
    ),
    actionButton(
      inputId = "inp2",
      label = "button 2"
    ),

    actionButton(
      inputId = "inp3",
      label = "button 3"
    ) \%>\%
      describe_using(descID = "descriptor"), # it's happening here

    actionButton(
      inputId = "inp4",
      label = "button 4"
    ),
    actionButton(
      inputId = "inp5",
      label = "button 5"
    ),

    div(
      id = "descriptor",
      "button 3 is just amazing! click on it!!!"
    )
  )

  server <- function(input, output, session) {}

  shinyApp(ui, server)
}
}
