% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tabbable.R
\name{make_tabable}
\alias{make_tabable}
\title{Make HTML elements tabable}
\usage{
make_tabable(elements, tab_index = 0)
}
\arguments{
\item{elements}{the HTML elements to be tabable (if not by default)}

\item{tab_index}{takes either 0, a negative or a positive value according to the required state of the element.
0 will make the element tabable with its relative order defined by the platform convention.
a negative value will make the element untabable.
a positive value will make the element tabable and its relative order defined by the provided value.}
}
\value{
a tabable HTML element
}
\description{
Make HTML elements tabable
}
